+++ src_cuda/gemv_pp.cu
@@ -95,7 +95,7 @@ void internal::Cholesky::gemv_pp_f64(cudaStream_t stream, int32_t j, int32_t M, 
   int32_t grid_n = std::min(grid_blocks, (N + block_threads - 1) / block_threads);
   int32_t grid = std::max(grid_n, std::min(grid_blocks, (M + block_threads - 1) / block_threads));
   gemv_pp_kernel<double* __restrict__, double* __restrict__, double_idx, double_idx* __restrict__, grid_blocks, block_threads>
-    <<< grid, block_threads, 0, stream >>> (j, M, N, sq[0], sq[1], A, lda, D, (double_idx*)sq);
+    <<< grid, block_threads, 0, stream >>> (j, M, N, sq[0], sq[1], A, int64_t(lda), D, (double_idx*)sq);
   imax_f64_host_sync(stream, N, grid_n, sq);
 }
 
@@ -103,7 +103,7 @@ void internal::Cholesky::gemv_pp_f32(cudaStream_t stream, int32_t j, int32_t M, 
   int32_t grid_n = std::min(grid_blocks, (N + block_threads - 1) / block_threads);
   int32_t grid = std::max(grid_n, std::min(grid_blocks, (M + block_threads - 1) / block_threads));
   gemv_pp_kernel<float* __restrict__, float* __restrict__, float_idx, float_idx* __restrict__, grid_blocks, block_threads>
-    <<< grid, block_threads, 0, stream >>> (j, M, N, sq[0], sq[1], A, lda, D, (float_idx*)sq);
+    <<< grid, block_threads, 0, stream >>> (j, M, N, sq[0], sq[1], A, int64_t(lda), D, (float_idx*)sq);
   imax_f32_host_sync(stream, N, grid_n, sq);
 }
 
@@ -111,7 +111,7 @@ void internal::Cholesky::gemv_pp_f128_dd(cudaStream_t stream, int32_t j, int32_t
   int32_t grid_n = std::min(grid_blocks, (N + block_threads - 1) / block_threads);
   int32_t grid = std::max(grid_n, std::min(grid_blocks, (M + block_threads - 1) / block_threads));
   gemv_pp_kernel<double2* __restrict__, double2* __restrict__, double2_idx, double2_idx* __restrict__, grid_blocks, block_threads>
-    <<< grid, block_threads, 0, stream >>> (j, M, N, sq[0], sq[1], A, lda, D, (double2_idx*)sq);
+    <<< grid, block_threads, 0, stream >>> (j, M, N, sq[0], sq[1], A, int64_t(lda), D, (double2_idx*)sq);
   imax_f128_dd_host_sync(stream, N, grid_n, sq);
 }
 
@@ -119,7 +119,7 @@ void internal::Cholesky::gemv_pp_f128_qf(cudaStream_t stream, int32_t j, int32_t
   int32_t grid_n = std::min(grid_blocks, (N + block_threads - 1) / block_threads);
   int32_t grid = std::max(grid_n, std::min(grid_blocks, (M + block_threads - 1) / block_threads));
   gemv_pp_kernel<float4* __restrict__, float4* __restrict__, float4_idx, float4_idx* __restrict__, grid_blocks, block_threads>
-    <<< grid, block_threads, 0, stream >>> (j, M, N, sq[0], sq[1], A, lda, D, (float4_idx*)sq);
+    <<< grid, block_threads, 0, stream >>> (j, M, N, sq[0], sq[1], A, int64_t(lda), D, (float4_idx*)sq);
   imax_f128_qf_host_sync(stream, N, grid_n, sq);
 }
 
@@ -128,7 +128,7 @@ void internal::Cholesky::gemv_pp_cf64(cudaStream_t stream, int32_t j, int32_t M,
   int32_t grid = std::max(grid_n, std::min(grid_blocks, (M + block_threads - 1) / block_threads));
   cuDoubleComplex sqc = make_cuDoubleComplex(*sq, 0.);
   gemv_pp_kernel<double* __restrict__, cuDoubleComplex* __restrict__, double_idx, double_idx* __restrict__, grid_blocks, block_threads>
-    <<< grid, block_threads, 0, stream >>> (j, M, N, sqc, sq[1], (cuDoubleComplex*)A, lda, D, (double_idx*)sq);
+    <<< grid, block_threads, 0, stream >>> (j, M, N, sqc, sq[1], (cuDoubleComplex*)A, int64_t(lda), D, (double_idx*)sq);
   imax_f64_host_sync(stream, N, grid_n, sq);
 }
 
@@ -137,7 +137,7 @@ void internal::Cholesky::gemv_pp_cf32(cudaStream_t stream, int32_t j, int32_t M,
   int32_t grid = std::max(grid_n, std::min(grid_blocks, (M + block_threads - 1) / block_threads));
   cuComplex sqc = make_cuComplex(*sq, 0.f);
   gemv_pp_kernel<float* __restrict__, cuComplex* __restrict__, float_idx, float_idx* __restrict__, grid_blocks, block_threads>
-    <<< grid, block_threads, 0, stream >>> (j, M, N, sqc, sq[1], (cuComplex*)A, lda, D, (float_idx*)sq);
+    <<< grid, block_threads, 0, stream >>> (j, M, N, sqc, sq[1], (cuComplex*)A, int64_t(lda), D, (float_idx*)sq);
   imax_f32_host_sync(stream, N, grid_n, sq);
 }
 
@@ -146,7 +146,7 @@ void internal::Cholesky::gemv_pp_cf128_dd(cudaStream_t stream, int32_t j, int32_
   int32_t grid = std::max(grid_n, std::min(grid_blocks, (M + block_threads - 1) / block_threads));
   complex_double2 sqc = device::dd::make_complex_double2(*sq, make_double2(0., 0.));
   gemv_pp_kernel<double2* __restrict__, complex_double2* __restrict__, double2_idx, double2_idx* __restrict__, grid_blocks, block_threads>
-    <<< grid, block_threads, 0, stream >>> (j, M, N, sqc, sq[1], A, lda, D, (double2_idx*)sq);
+    <<< grid, block_threads, 0, stream >>> (j, M, N, sqc, sq[1], A, int64_t(lda), D, (double2_idx*)sq);
   imax_f128_dd_host_sync(stream, N, grid_n, sq);
 }
 
@@ -155,6 +155,6 @@ void internal::Cholesky::gemv_pp_cf128_qf(cudaStream_t stream, int32_t j, int32_
   int32_t grid = std::max(grid_n, std::min(grid_blocks, (M + block_threads - 1) / block_threads));
   complex_float4 sqc = device::qf::make_complex_float4(*sq, make_float4(0.f, 0.f, 0.f, 0.f));
   gemv_pp_kernel<float4* __restrict__, complex_float4* __restrict__, float4_idx, float4_idx* __restrict__, grid_blocks, block_threads>
-    <<< grid, block_threads, 0, stream >>> (j, M, N, sqc, sq[1], A, lda, D, (float4_idx*)sq);
+    <<< grid, block_threads, 0, stream >>> (j, M, N, sqc, sq[1], A, int64_t(lda), D, (float4_idx*)sq);
   imax_f128_qf_host_sync(stream, N, grid_n, sq);
 }
